requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
jinja2
python-dotenv
google-genai
google-generativeai
python-multipart
markdown